ethernet.{ch}: update type safety of ss_frame_handle

+++ src/ethernet.c
@@ -24,7 +24,7 @@
 #include "sdn_sensor.h"
 #include "sensor_conf.h"
 
-void ss_frame_handle(rte_mbuf_t* mbuf, unsigned int lcore_id, uint8_t port_id) {
+void ss_frame_handle(rte_mbuf_t* mbuf, uint16_t lcore_id, uint8_t port_id) {
     int rv;
     ss_frame_t rx_buf;
     ss_frame_t tx_buf;
